Add DuplicateTable and duplicate finding in DirTree
We add module for DuplicateTable - the data structure responsible for
finding duplicates for each file. In the module the files are added to a
hash map with the checksums serving as keys in the hash map. We get the
duplicates as entries by getting all the values corresponding to the
particular checksum.