Fix getting parent node data in DirTree
There was a bug that when there was too big a difference between the
levels of two duplicate nodes, we could get a panic because we were
searching for a parent of a root node or root dir.

The function _get_parent_table_data now returns a Result to account for
that case - we return Err if the child data is that of root dir or node.