Add threads to duplicate table
In order to speed up checksum calculation we add threading to duplicate
table and calculate checksums in separate threads.

When testing the change, we have found that when used with slower
external hdds multithreading can paradoxically slow down the overall
progress, so we set single threading as a default for now.