Refactor adding duplicates to duplicate list
The previous method of checking if any of the paths in duplicate list
contains any of the paths added was too expensive, so it is replaced.

The new method adds flags to each node marking whether a node is a child
node of already added duplicate, parent of already added duplicate,
duplicate itself or none of the above.

When we add a node to the duplicate list, we first check the flag on the
node. If it is marked as a child of duplicate we do nothing, otherwise
we add the corresponding duplicate object to the list. If it is marked
as parent of duplicate, we remove the duplicate object corresponding to
the descendant of the node we are adding.