Add progress bar
We add progress bar to the DuDe. In order to avoid unnecessary
dependencies in the library we add a two traits - ProgressIndicator for
simple progress bars and ProgressMultiline for displaying both overall
progress of getting file duplicates and directory processing at the same
time. We use trait objects with these two traits in the library. This
enables the user of the library to write his own (or just copy mine)
implementation of the progress bars.

The default trait objects used in the library are shell objects -
NoProgressIndicator and NoProgressMultiline - that display nothing.

For the CLI application we implement progress bars and spinners using
indicatif crate.